Tweak exporting default at the end of files
