+++ src/index.ts
@@ -1,7 +1,8 @@
 import OptionPrefix from './@types/OptionPrefix';
 import MessageOption from './message-option';
 import OptionDef from './option-def';
-import { VCommandParser, VLazyParsedCommand, VParsedCommand } from './vcommandparser';
+import VCommandParser from './vcommandparser';
+import { VLazyParsedCommand, VParsedCommand } from './vparsedcommand';
 
 export {
 	VCommandParser,
